requirements.txt
exifread==3.0.0
Pillow==9.3.0
python-dotenv==0.21.0
simplejson==3.18.0
